+++ scrapers/cnn_scraper.py
@@ -5,7 +5,9 @@ this module contains helper functions to scrape CNN
 from urllib.parse import urljoin
 from bs4 import Tag
 import yaml
-from scrape_helper import make_request
+
+from scrapers.scrape_helper import make_request
+from models.data_models import Article
 
 
 def get_top_news(category: str, limit: int = 3) -> list[dict[str, str]]:
@@ -21,7 +23,7 @@ def get_top_news(category: str, limit: int = 3) -> list[dict[str, str]]:
     a hyperlink reference
     """
 
-    with open("cnn_config.yaml", "r", encoding="UTF-8") as config_file:
+    with open("config/cnn_config.yaml", "r", encoding="UTF-8") as config_file:
         cnn_config = yaml.safe_load(config_file)
 
     section_details = cnn_config["sections"][category]
@@ -60,7 +62,7 @@ def get_article_text(path: str) -> list[str]:
     the scraping logic is that it seems like all pargraphs from the articles come
     from paragraph attributes with parent of class 'article__cnontent'.
     """
-    with open("cnn_config.yaml", "r", encoding="UTF-8") as config_file:
+    with open("config/cnn_config.yaml", "r", encoding="UTF-8") as config_file:
         cnn_config = yaml.safe_load(config_file)
 
     base_url = cnn_config["base_url"]
@@ -78,3 +80,21 @@ def get_article_text(path: str) -> list[str]:
         paragraph_list = [p_obj.text.strip() for p_obj in article_content.find_all(name="p")]
 
     return paragraph_list
+
+
+def get_articles(category: str, limit: int = 3) -> list[Article]:
+    """
+    returns a list of Article objects storing path, title, and paragraph texts from the
+    given category, with length no more than limit.
+
+    category: category within cnn to scrape articles from
+    limit: the maximum number of articles to scrape
+    """
+    article_list = []
+
+    top_news_list = get_top_news(category, limit)
+    for news in top_news_list:
+        paragraphs = get_article_text(news["path"])
+        article_list.append(Article(path=news["path"], title=news["title"], text=paragraphs))
+
+    return article_list
